spank plugin: revert rank to a signed integer
In an earlier commit, the rank option in the io-watchdog spank plugin
was inadvertently changed to an unsigned int. This invalidated the
comparison of rank < 0 to determine if the user had requested a rank
other than the default, which causes most uses of the io-watchdog
SLURM plugin to silently fail.

This patch reverts the rank variable to a signed integer.

Fixes CHAOS bz#1031

don't need rank to be long